requirements.txt
fastapi
uvicorn
websockets
pymysql
sqlalchemy